+++ py-polars/polars/io/excel/__init__.py
@@ -1,0 +1,3 @@
+from polars.io.excel.functions import read_excel
+
+__all__ = ["read_excel"]
